Refactoring: add constant for default name of sitemap
The plan is to make this important value more visible
before changing it from "sitemap_index.xml" to "sitemap.xml".

It would be nice to have this value synchronized with the default
prefix of sitemap files, but since I found no easy way to do it,
this is best left to the reader.