Add assertions on contents of generated sitemap indexes in unit tests
The new unit tests rely on XPath expressions, which require the
declaration of the namespace "http://www.sitemaps.org/schemas/sitemap/0.9"
used for sitemaps and sitemap indexes. Sadly, there is no implementation
of the NamespaceContext interface readily available in Java 6. I added the
Apache WebServices Common Utilities to dependencies of the tests to use
the NamespaceContextImpl class part of this package.

In order to evaluate multiple XPath expressions on the same generated
sitemap index, a DOM node must be constructed first. I added a utility
method in TestUtil class that reads the whole file as a String and returns
the document Node for further evaluations.